+++ CHANGES.rst
@@ -1,4 +1,11 @@
 
+0.2.0
+---------------------
+
+* Add audio file rename options.
+* Fix no arguments raising parsing error. Know does default ``--help``.
+* Fix runtime execution path not found to metadata.
+
 0.1.0
 ---------------------